backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
django-cors-headers==4.9.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
